pyproject.toml
[project]
name = "promptlabs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-allauth>=65.13.1",
    "djangorestframework>=3.16.1",
    "pillow>=12.1.0",
]
